+++ noir/noir-repo/compiler/noirc_evaluator/src/brillig/brillig_ir/codegen_control_flow.rs
@@ -148,7 +148,7 @@ impl BrilligContext {
         condition: SingleAddrVariable,
         revert_data_items: Vec<BrilligVariable>,
         revert_data_types: Vec<BrilligParameter>,
-        revert_data_type: usize,
+        revert_data_type: u64,
     ) {
         assert!(condition.bit_size == 1);
 
@@ -162,7 +162,8 @@ impl BrilligContext {
 
             let current_revert_data_pointer = ctx.allocate_register();
             ctx.mov_instruction(current_revert_data_pointer, revert_data.pointer);
-            let revert_data_id = ctx.make_usize_constant_instruction(revert_data_type.into());
+            let revert_data_id =
+                ctx.make_usize_constant_instruction((revert_data_type as u128).into());
             ctx.store_instruction(current_revert_data_pointer, revert_data_id.address);
             ctx.codegen_usize_op_in_place(current_revert_data_pointer, BrilligBinaryOp::Add, 1);
             for (revert_variable, revert_param) in
